+++ sdks/react-native/react-native.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  dependency: {
+    platforms: {
+      android: {
+        packageImportPath: "import com.autlantic.checkout.AutlanticCheckoutPackage;",
+        packageInstance: "new AutlanticCheckoutPackage()",
+      },
+      ios: {},
+    },
+  },
+};